image_service/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
drf-spectacular
drf-spectacular-sidecar
kafka-python
faker>=18.11.2
django-filter
elasticsearch>=7.0.0,<8.0.0
Pillow>=9.0.0
requests>=2.25.0
whitenoise>=6.0.0
dj-database-url
boto3>=1.28.0
django-storages>=1.14.0 